Make confluence_version a First Class Fact
This stems from the fact that the confluence module unconditionally manages the
directory /etc/puppetlabs/facter/facts.d directory in the sense that if the code
gets evaluated first you run into the age old race condition of clashing with other
classes who wish to manage it.  This removes said management, moves the fact
to a ruby/puppet native which yields slightly prettier and less code.

The change to `confluence::service` was required to get CI to pass.  I'd properly
refactor this but it'd probably destroy users who already reference in hiera
`confluence::service::service_file_location` etc.