Make confluence_version a First Class Fact
This stems from the fact that the confluence module unconditionally manages the
directory /etc/puppetlabs/facter/facts.d directory in the sense that if the code
gets evaluated first you run into the age old race condition of clashing with other
classes who wish to manage it.  This removes said management, moves the fact
to a ruby/puppet native which yields slightly prettier and less code.  Transitional
code has been put in place to remove the old fact.

The `confluence::service` class is dependant on the old inherits which used to live
in `confluence::facts` to get access to `confluence::params`.  I've moved this to
the more socially acceptable top level class.