Allow System Group
When managing the confluence user it defaults to a system user, however this doesn't
apply to the group which may be problematic in deployments where other groups are
hard coded to 1000 which has already been taken by confluence.  Make the confluence
group a system account by default, this appears not to alter existing installations.