Use 'ps' for version fact
The long arguments to `pgrep` are not present in older versions of
pgrep, as provided by 'procps'. procps-ng appears to have introduced
these, as used by EL7.

This commit changes the version fact to use `ps ax | grep` to find the
process. It's generic enough and should be supported by the platforms
listed in the module's support matrix.