Update setenv.sh template
The setenv.sh template, while functional, appeared to be a bit old and
lacking some updated conventions Atlassian's been using in the setenv.sh
script for at least a couple of years (such as CATALINA_OPTS).

Most importantly, there's some new additions to it to support Confluence
6 and its synchrony service.

This preserves the `JAVA_OPTS` to be backwards compatible, but does
move to use `CATALINA_OPTS`, as has been used by Atlassian out of the
box for some time, and introduces a `catalina_opts` parameter.